Install superagent and setup basic integration test environment
