+++ package.json
@@ -40,7 +40,8 @@
     "nodemon": "1.12.1",
     "prettier": "1.8.2",
     "prettier-eslint": "8.2.2",
-    "prettier-eslint-cli": "4.4.0"
+    "prettier-eslint-cli": "4.4.0",
+    "supertest": "^3.0.0"
   },
   "scripts": {
     "precommit": "lint-staged",
