+++ src/resolvers/questionInstances.integration.test.js
@@ -1,0 +1,10 @@
+const request = require('supertest')
+
+const server = require('../app')
+
+describe('questionInstances', () => {
+  it.skip('works', async () => {
+    const response = await request(server).post('/graphql', {})
+    expect(response).toMatchSnapshot()
+  })
+})
